Refactor to use new UCD framework
