+++ java/drivers/driver-hazelcast4plus/src/main/java/com/hazelcast/simulator/tests/ucd/topic/classes/CompletableMessageListener.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 2008-2016, Hazelcast, Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.hazelcast.simulator.tests.ucd.topic.classes;
+
+import com.hazelcast.topic.Message;
+import com.hazelcast.topic.MessageListener;
+
+import java.util.concurrent.CompletableFuture;
+
+public class CompletableMessageListener<T> implements MessageListener<T> {
+    private final CompletableFuture<T> future;
+
+    public CompletableMessageListener(CompletableFuture<T> future) {
+        this.future = future;
+    }
+
+    @Override
+    public void onMessage(Message<T> message) {
+        future.complete(message.getMessageObject());
+    }
+}
